Invoice_Data_Extraction/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
pymupdf
requests
langchain
openai
python-dotenv
pdf2image